Regroup github handlers ITs behind an abstract GitHubEventHandlerIT
